apply changes to precommit hook and reorganise imports
