+++ rustfmt.toml
@@ -1,8 +1,8 @@
 comment_width = 100
 format_code_in_doc_comments = true
 wrap_comments = true
-reorder_imports = false
+reorder_imports = true
 edition = "2021"
-imports_granularity = "Crate"
-group_imports = "Preserve"
+imports_granularity = "Module"
+group_imports = "StdExternalCrate"
 style_edition = "2021"
